modif script database et les entities

+++ src/main/java/fr/dta/mediatic/utils/model/AbstractModel.java
@@ -10,13 +10,13 @@ public abstract class AbstractModel {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
+	private int id;
 	
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 }
